requirements.txt
flask>=2.0.0
openai>=0.27.0
azure-search-documents>=11.4.0
gunicorn>=20.1.0
uvicorn>=0.17.0
requests>=2.28.0
bcrypt>=3.2.0 --only-binary=:all:  # Add this line
python-dotenv>=0.19.0  # Add if using .env files